# api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "De Mol API",
    "version": "v1"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/api/games": {
      "post": {
        "tags": [
          "Games"
        ],
        "operationId": "CreateGame",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGameRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Game"
                }
              }
            }
          }
        }
      }
    },
    "/api/games/{gameId}": {
      "get": {
        "tags": [
          "Games"
        ],
        "operationId": "GetGame",
        "parameters": [
          {
            "name": "gameId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Game"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Games"
        ],
        "operationId": "DeleteGame",
        "parameters": [
          {
            "name": "gameId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/games/{gameId}/join": {
      "post": {
        "tags": [
          "Games"
        ],
        "operationId": "JoinGame",
        "parameters": [
          {
            "name": "gameId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinGameRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/games/by-invite/{inviteCode}": {
      "get": {
        "tags": [
          "Games"
        ],
        "operationId": "GetGameByInvite",
        "parameters": [
          {
            "name": "inviteCode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameSummaryResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/games/{gameId}/contestants": {
      "post": {
        "tags": [
          "Games"
        ],
        "operationId": "AddContestants",
        "parameters": [
          {
            "name": "gameId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddContestantsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Game"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddContestantsRequest": {
        "required": [
          "contestants"
        ],
        "type": "object",
        "properties": {
          "contestants": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/Contestant"
            }
          }
        }
      },
...
      "Contestant": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "age": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "photoUrl": {
            "type": "string"
          },
          "eliminatedInEpisode": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "CreateGameRequest": {
        "required": [
          "name",
          "contestants"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": [
              "null",
              "string"
            ]
          },
          "contestants": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/Contestant"
            }
          }
        }
      },
      "Episode": {
        "type": "object",
        "properties": {
          "number": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "deadline": {
            "type": "string",
            "format": "date-time"
          },
          "eliminatedContestantIds": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "Game": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "adminUserId": {
            "type": "string"
          },
          "contestants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Contestant"
            }
          },
          "episodes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Episode"
            }
          },
          "moleContestantId": {
            "type": [
              "null",
              "string"
            ]
          },
          "inviteCode": {
            "type": "string"
          }
        }
      },
      "GameSummaryResponse": {
        "required": [
          "id",
          "name",
          "contestantCount",
          "episodeCount"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "contestantCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "episodeCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "JoinGameRequest": {
        "required": [
          "inviteCode"
        ],
        "type": "object",
        "properties": {
          "inviteCode": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "MessageResponse": {
        "required": [
          "message"
        ],
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}